Move some IP lists functions to lib/
