+++ src/spind/spind.h
@@ -11,9 +11,6 @@
 void maybe_sendflow(flow_list_t *flow_list, time_t now);
 void report_block(int af, int proto, uint8_t *src_addr, uint8_t *dest_addr, unsigned src_port, unsigned dest_port, int payloadsize);
 
-// Move these to, say, ip_list.[ch]
-int ip_in_ignore_list(ip_t* ip);
-int addr_in_ignore_list(int family, uint8_t* addr);
 void send_command_dnsquery(dns_pkt_info_t* pkt_info);
 
 // RPC section
